requirements.txt
bcrypt==3.2.0
cffi==1.14.4
cryptography==3.3.1
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
paramiko==2.7.2
picosdk==1.0
Pillow==8.0.1
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
tqdm==4.54.1
